+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -470,36 +470,6 @@ class IPrenotazioniFolder(model.Schema):
         default=[],
     )
 
-    how_to_get_here = schema.Text(
-        required=False,
-        title=_("How to get here", default="How to get here"),
-        description=_("Insert here indications on how to reach the office"),
-    )
-
-    phone = schema.TextLine(
-        title=_("Contact phone"),
-        description=_("Insert here the contact phone"),
-        required=False,
-    )
-
-    fax = schema.TextLine(
-        title=_("Contact fax"),
-        description=_("Insert here the contact fax"),
-        required=False,
-    )
-
-    pec = schema.TextLine(
-        title=_("Contact PEC"),
-        description=_("Insert here the contact PEC"),
-        required=False,
-    )
-
-    complete_address = schema.Text(
-        required=False,
-        title=_("Complete address", default="Complete address"),
-        description=_("Insert here the complete office address"),
-    )
-
     @invariant
     def data_validation(data):
         """
@@ -670,16 +640,6 @@ class IPrenotazioniFolder(model.Schema):
         ],
     )
 
-    form.omitted("how_to_get_here", "phone", "fax", "pec", "complete_address")
-    model.fieldset(
-        "contacts",
-        label=_("contacts_label", default="Contacts"),
-        description=_(
-            "contacts_help",
-            default="Show here contacts information that will be used by authomatic mail system",  # noqa
-        ),
-        fields=["how_to_get_here", "phone", "fax", "pec", "complete_address"],
-    )
     model.fieldset(
         "Notifications",
         label=_("notifications_label", default="Notifications"),